requirements.txt
# MCP Server
mcp>=1.0.0

# AI Agent
anthropic>=0.39.0

# Environment variables
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
